update stage locator to theia version and results cleanup
the stage locator as it was did not work with the theia datasets, so I merged my refactored and debugged version with the theia version found in the previous runs (artifact from original repo). it should now work with trace as well but it might require to comment in some of the lines. did not test compatibility.